+++ tests/unittests/apps/test1/model.py
@@ -1,0 +1,5 @@
+
+import module1
+
+def app():
+    return 'Testing test1 module import: ' + module1.get_text()