+++ tests/unittests/apps/test1/test_module.py
@@ -1,0 +1,2 @@
+def get_text():
+    return "text for test1"