+++ tests/unittests/apps/test2/module1.py
@@ -1,0 +1,5 @@
+
+import module2
+
+def get_text():
+    return module2.get_text()