+++ tests/unittests/apps/test2/module2.py
@@ -1,0 +1,3 @@
+
+def get_text():
+    return 'text for test2'