+++ tests/unittests/apps/test2/test_module.py
@@ -1,0 +1,5 @@
+import logging
+
+def get_text():
+    logging.debug('text for test2')
+    return "text for test2"