+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [6.21.0] - 2023-05-01
 - add quiet logging completion of ajax requests
 - update admin overview to use quiet logging
 - add support for multiple groups in User.is_member
@@ -13,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - produce page title when search is empty
 - use content template by default for content pages
 - fix images manager of content app
+- refactor module loading for Python 3.9
+- promote locate_user to top level function
 
 ## [6.20.0] - 2022-05-27
 - refactor and enhance icons app